Pc branch

+++ project/Models/Gallon.py
@@ -2,4 +2,4 @@ class Gallon:
     quart_const = 4
 
     def __init__(self, amount):
-        self.amount = float(amount) * self.quart_const
+        self.amount = int(amount) * self.quart_const
