+++ project/Models/Quart.py
@@ -1,3 +1,3 @@
 class Quart:
     def __init__(self, amount):
-        self.amount = float(amount)
+        self.amount = int(amount)
